change int to double

+++ src/MySpaceMatrix.java
@@ -2,22 +2,22 @@ import java.util.ArrayList;
 
 public class MySpaceMatrix {
 
-    ArrayList<int[]> ds2;
-    ArrayList<int[]> ds3;
+    ArrayList<double[]> ds2;
+    ArrayList<double[]> ds3;
 
-    public MySpaceMatrix(int[][] normalMatrix) {
+    public MySpaceMatrix(double[][] normalMatrix) {
         this.ds2 = changeNormalMatrixToListOfRows(normalMatrix);
         this.ds3 = changeNormalMatrixToListOfColumns(normalMatrix);
     }
 
 
-    private static ArrayList<int[]> changeNormalMatrixToListOfRows(int[][] normalMatrix) {
-        ArrayList<int[]> listOfRows = new ArrayList<>();
+    private static ArrayList<double[]> changeNormalMatrixToListOfRows(double[][] normalMatrix) {
+        ArrayList<double[]> listOfRows = new ArrayList<>();
 
-        for (int[] matrix : normalMatrix) {
+        for (double[] matrix : normalMatrix) {
             for (int j = 0; j < normalMatrix[0].length; j++) {
                 if (matrix[j] != 0) {
-                    int[] element = {j, matrix[j]};
+                    double[] element = {j, matrix[j]};
                     listOfRows.add(element);
                 }
             }
@@ -25,13 +25,13 @@ public class MySpaceMatrix {
         return listOfRows;
     }
 
-    private static ArrayList<int[]> changeNormalMatrixToListOfColumns(int[][] normalMatrix) {
-        ArrayList<int[]> listOfRows = new ArrayList<>();
+    private static ArrayList<double[]> changeNormalMatrixToListOfColumns(double[][] normalMatrix) {
+        ArrayList<double[]> listOfRows = new ArrayList<>();
 
         for (int i = 0; i < normalMatrix.length; i++) {
             for (int j = 0; j < normalMatrix[i].length; j++) {
                 if (normalMatrix[i][j] != 0) {
-                    int[] element = {i, normalMatrix[i][j]};
+                    double[] element = {i, normalMatrix[i][j]};
                     listOfRows.add(element);
                 }
             }
